postgres/requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
psycopg2-binary==2.7.7
PyYAML==3.13
requests==2.21.0
urllib3==1.24.1
